Move instance domain files to correct subfolders
