+++ src/domain/metadata/usecases/MetadataSyncUseCase.ts
@@ -7,8 +7,8 @@ import { D2Api } from "../../../types/d2-api";
 import { ExportBuilder, NestedRules, SynchronizationBuilder } from "../../../types/synchronization";
 import { promiseMap } from "../../../utils/common";
 import { Ref } from "../../common/entities/Schemas";
-import InstanceEntity from "../../instance/Instance";
-import InstanceRepository from "../../instance/InstanceRepository";
+import InstanceEntity from "../../instance/entities/Instance";
+import InstanceRepository from "../../instance/repositories/InstanceRepository";
 import { SynchronizationResult } from "../../synchronization/entities/SynchronizationResult";
 import {
     GenericSyncUseCase,
